test: split golden test files into separate fixtures directory
