+++ cmd/lx/testdata/languages/Main.java
@@ -1,0 +1,16 @@
+public class Calculator {
+    private int value;
+    public static final int MAX = 100;
+
+    public Calculator(int initial) {
+        this.value = initial;
+    }
+
+    public int add(int x) {
+        return value + x;
+    }
+
+    private int secret() {
+        return 42;
+    }
+}
