+++ cmd/lx/testdata/languages/main.c
@@ -1,0 +1,10 @@
+typedef struct Point {
+    int x;
+    int y;
+} Point;
+
+int add(int a, int b) {
+    return a + b;
+}
+
+void greet(const char *name);
