+++ cmd/lx/testdata/languages/main.cpp
@@ -1,0 +1,14 @@
+class Widget {
+public:
+    int value;
+    void set(int v) {
+        value = v;
+    }
+private:
+    int secret;
+    void hide();
+};
+
+int FreeFn(int v) {
+    return v;
+}
