+++ cmd/lx/testdata/languages/main.js
@@ -1,0 +1,11 @@
+export class Greeter {
+    message = "hi";
+
+    greet(name) {
+        return this.message + " " + name;
+    }
+}
+
+export function topLevel(x) {
+    return x + 1;
+}
